config: use structures from phraseapp-go library
Now the configuration types are used from the phraseapp-go library to
enable configuration validation and remove unnecessary config parse
runs.

* config.go: Remove all the custom types and parse functions. There now
  is a single ReadConfig method that returns the libraries config type.
* main.go: Report errors in configuration and send configuration to the
  router so it can read required parameters.
* actions (pull, push and wizard): Use the config.

The credentials are found in the configuration. They are first read from
the configuration and than overwritten by the router if necessary.

For push and pull the source and target types are unmarshalled from yaml
into some map, marshalled back to yaml and finally marshalled to the
respective type here. This is required as the actual implementation for
those types shouldn't reside in the phraseapp go library.