config: remove usage of user.Current
user.Current uses cgo which has disastrous consequences when
cross-compiling (okay the call will return with nothing). This made the
user global config file a darwin (because we build here) only feature.
Now it uses the HOME or HomePath environment variable to determine the
location of the default config file.