tests/push: fix issue
This is related to the change in getRemoteLocaleForLocaleFile that uses
the remote's code instead of name to match against the locale-file's
code.

I'm currently not sure this is intended behavior: as a remote locale
with different name, but matching code would be used. And if multiple
remote locales with the same code exist, the first would always be
choosen (even if a later one would be a better fit).