push: fix remote locale matching
The remote locale matcher tries to determine the remote locale to be
used for upload for a given file. This was changed to take into account
all specified values, i.e. for a file that specifies code "en" and name
"american", a remote locale with code "en" and name "english" would
_not_ be taken into account.

If multiple candidates match the specified values the first is returned
at the moment. This is error prone and should maybe changed to return an
error!