Make Setup recognize Windows Version and throw error
The setup is now scoped on x64 systems only as our client currently only
works on x64 systems, not on x86 systems.

This way the users gets to know right away, that the CLI is not
compatible with their windows version. Otherwise, the user could install
our CLI just to find it not working.